modified:   src/main/java/com/example/hijra_kyc/HijraKycApplication.java 	modified:   src/main/java/com/example/hijra_kyc/controller/BackReasonController.java 	modified:   src/main/java/com/example/hijra_kyc/controller/BranchController.java 	modified:   src/main/java/com/example/hijra_kyc/controller/ImageController.java 	modified:   src/main/java/com/example/hijra_kyc/controller/MakeFormController.java 	modified:   src/main/java/com/example/hijra_kyc/controller/MessageController.java 	modified:   src/main/java/com/example/hijra_kyc/controller/RoleController.java 	modified:   src/main/java/com/example/hijra_kyc/controller/SystemLogController.java 	modified:   src/main/java/com/example/hijra_kyc/controller/UserProfileController.java 	modified:   src/main/java/com/example/hijra_kyc/dto/BackReasonInDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/BackReasonOutDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/BranchInDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/BranchOutDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/ImageDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/ImageReturnDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/MakeFormDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/MakeFormOutDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/MessageInDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/MessageOutDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/PermissionInDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/PermissionOutDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/RoleInDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/RoleOutDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/RolePermissionInDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/RolePermissionOutDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/SystemLogInDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/SystemLogOutDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/UserProfileInDto.java 	modified:   src/main/java/com/example/hijra_kyc/dto/UserProfileOutDto.java 	modified:   src/main/java/com/example/hijra_kyc/exception/GlobalException.java 	modified:   src/main/java/com/example/hijra_kyc/mapper/BackReasonMapper.java 	modified:   src/main/java/com/example/hijra_kyc/mapper/BranchMapper.java 	modified:   src/main/java/com/example/hijra_kyc/mapper/ImageMapper.java 	modified:   src/main/java/com/example/hijra_kyc/mapper/MakeFormMapper.java 	modified:   src/main/java/com/example/hijra_kyc/mapper/PermissionMapper.java 	modified:   src/main/java/com/example/hijra_kyc/mapper/RoleMapper.java 	modified:   src/main/java/com/example/hijra_kyc/mapper/RolePermissionMapper.java 	modified:   src/main/java/com/example/hijra_kyc/mapper/SystemLogMapper.java 	modified:   src/main/java/com/example/hijra_kyc/mapper/UserProfileMapper.java 	modified:   src/main/java/com/example/hijra_kyc/model/BackReason.java 	modified:   src/main/java/com/example/hijra_kyc/model/Branch.java 	modified:   src/main/java/com/example/hijra_kyc/model/Image.java 	modified:   src/main/java/com/example/hijra_kyc/model/MakeForm.java 	modified:   src/main/java/com/example/hijra_kyc/model/Message.java 	modified:   src/main/java/com/example/hijra_kyc/model/Permission.java 	modified:   src/main/java/com/example/hijra_kyc/model/Role.java 	modified:   src/main/java/com/example/hijra_kyc/model/SystemLog.java 	modified:   src/main/java/com/example/hijra_kyc/model/UserProfile.java 	modified:   src/main/java/com/example/hijra_kyc/repository/BackReasonRepository.java 	modified:   src/main/java/com/example/hijra_kyc/repository/BranchRepository.java 	modified:   src/main/java/com/example/hijra_kyc/repository/ImageRepository.java 	modified:   src/main/java/com/example/hijra_kyc/repository/MakeFormRepository.java 	modified:   src/main/java/com/example/hijra_kyc/repository/MessageRepository.java 	modified:   src/main/java/com/example/hijra_kyc/repository/PermissionRepository.java 	modified:   src/main/java/com/example/hijra_kyc/repository/RolePermissionRepository.java 	modified:   src/main/java/com/example/hijra_kyc/repository/RoleRepository.java 	modified:   src/main/java/com/example/hijra_kyc/repository/SystemLogRepository.java 	modified:   src/main/java/com/example/hijra_kyc/repository/UserProfileRepository.java 	modified:   src/main/java/com/example/hijra_kyc/service/BackReasonService.java 	modified:   src/main/java/com/example/hijra_kyc/service/BranchService.java 	modified:   src/main/java/com/example/hijra_kyc/service/ImageService.java 	modified:   src/main/java/com/example/hijra_kyc/service/MakeFormService.java 	modified:   src/main/java/com/example/hijra_kyc/service/MessageService.java 	modified:   src/main/java/com/example/hijra_kyc/service/PermissionService.java 	modified:   src/main/java/com/example/hijra_kyc/service/RoleService.java 	modified:   src/main/java/com/example/hijra_kyc/service/SystemLogService.java 	modified:   src/main/java/com/example/hijra_kyc/service/UserProfileService.java 	modified:   src/main/resources/application.yml

+++ src/main/java/com/example/hijra_kyc/controller/MessageController.java
@@ -37,13 +37,13 @@ public class MessageController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteMessage(@PathVariable("id") int id, @RequestParam("senderId") int senderId){
+    public ResponseEntity<?> deleteMessage(@PathVariable("id") Long id, @RequestParam("senderId") Long senderId){
         Base<?> deleteMessage=messageService.deleteMessage(id, senderId);
         return baseService.rest(deleteMessage);
     }
 
     @PatchMapping("/messageEdit/{id}")
-    public ResponseEntity<?> editMessage(@RequestParam Long senderId, @RequestParam Long receiverId, @RequestBody MessageEdit message, @PathVariable("id") int id){
+    public ResponseEntity<?> editMessage(@RequestParam Long senderId, @RequestParam Long receiverId, @RequestBody MessageEdit message, @PathVariable("id") Long id){
         Base<?> messageEdit=messageService.updateMessage(senderId, receiverId,message, id);
         return baseService.rest(messageEdit);
     }
