+++ src/main/java/com/example/hijra_kyc/dto/SystemLogOutDto.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class SystemLogOutDto {
-    private int id;
-    private int userId;
+    private Long id;
+    private Long userId;
     private LocalDateTime actionTime;
     private String actionType;
 }
