+++ src/main/java/com/example/hijra_kyc/model/SystemLog.java
@@ -21,10 +21,11 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
+//TODO: make sure this model's functionality is finished at the backend , it should not extend to the frontend
 public class SystemLog {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @NotNull(message = "user id is required")
     @ManyToOne
     @JoinColumn(name = "user_id", referencedColumnName = "id", nullable = false)
