+++ src/main/java/com/example/hijra_kyc/repository/SystemLogRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.hijra_kyc.model.SystemLog;
 
-public interface SystemLogRepository extends JpaRepository<SystemLog, Integer>{
+public interface SystemLogRepository extends JpaRepository<SystemLog, Long>{
     
 }
