+++ src/main/java/com/example/hijra_kyc/repository/UserProfileRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.hijra_kyc.model.UserProfile;
 
-public interface UserProfileRepository extends JpaRepository<UserProfile, Integer>{
+public interface UserProfileRepository extends JpaRepository<UserProfile, Long>{
 
 }
